chore(mcp): move gmail to new architecture
