+++ front/lib/actions/mcp_internal_actions/servers/index.ts
@@ -16,7 +16,6 @@ import { default as databricksServer } from "@app/lib/actions/mcp_internal_actio
 import { default as deepDiveServer } from "@app/lib/actions/mcp_internal_actions/servers/deep_dive";
 import { default as freshserviceServer } from "@app/lib/actions/mcp_internal_actions/servers/freshservice";
 import { default as frontServer } from "@app/lib/actions/mcp_internal_actions/servers/front";
-import { default as gmailServer } from "@app/lib/actions/mcp_internal_actions/servers/gmail";
 import { default as driveServer } from "@app/lib/actions/mcp_internal_actions/servers/google_drive";
 import { default as httpClientServer } from "@app/lib/actions/mcp_internal_actions/servers/http_client";
 import { default as interactiveContentServer } from "@app/lib/actions/mcp_internal_actions/servers/interactive_content";
@@ -52,6 +51,7 @@ import { default as conversationFilesServer } from "@app/lib/api/actions/servers
 import { default as extractDataServer } from "@app/lib/api/actions/servers/extract_data";
 import { default as fileGenerationServer } from "@app/lib/api/actions/servers/file_generation";
 import { default as githubServer } from "@app/lib/api/actions/servers/github";
+import { default as gmailServer } from "@app/lib/api/actions/servers/gmail";
 import { default as calendarServer } from "@app/lib/api/actions/servers/google_calendar";
 import { default as sheetsServer } from "@app/lib/api/actions/servers/google_sheets";
 import { default as hubspotServer } from "@app/lib/api/actions/servers/hubspot";
