+++ front/lib/api/actions/servers/gmail/index.ts
@@ -1,0 +1,25 @@
+import type { McpServer } from "@modelcontextprotocol/sdk/server/mcp.js";
+
+import { makeInternalMCPServer } from "@app/lib/actions/mcp_internal_actions/utils";
+import { registerTool } from "@app/lib/actions/mcp_internal_actions/wrappers";
+import type { AgentLoopContextType } from "@app/lib/actions/types";
+import { GMAIL_TOOL_NAME } from "@app/lib/api/actions/servers/gmail/metadata";
+import { TOOLS } from "@app/lib/api/actions/servers/gmail/tools";
+import type { Authenticator } from "@app/lib/auth";
+
+function createServer(
+  auth: Authenticator,
+  agentLoopContext?: AgentLoopContextType
+): McpServer {
+  const server = makeInternalMCPServer("gmail");
+
+  for (const tool of TOOLS) {
+    registerTool(auth, agentLoopContext, server, tool, {
+      monitoringName: GMAIL_TOOL_NAME,
+    });
+  }
+
+  return server;
+}
+
+export default createServer;
